Uploaded file "Client.java"
This is a client to the TCP server. It is supposed to open a socket from a server, read the data from the server, and send data to the server. Finally it closes socket's connection to the server. Additional functionality will be added to this file, but for now it works as a baseline.